updated test-programm.py, added error.py
