+++ test-programm.py
@@ -1,4 +1,7 @@
 from main import *
+from error import *
+
+
 
 class Test:
 	def __init__(self):
@@ -14,22 +17,20 @@ class Test:
 			infoLog += self.error.getInfoString(info)+ "\n"
 		return errorLog + "\n" + infoLog
 	def test(self):
-		if normalize_url("/hallo.py.git") != "hallo.py":
-			self.infoLog.append("INF#001")
-		if github_repo_to_api("https://github.com/mitmirzutun") != "https://api.github.com/repos/mitmirzutun":
-			self.infoLog.append("INF#002")
-		if github_repo_to_api_releases("https://github.com/mitmirzutun") != "https://api.github.com/repos/mitmirzutun/releases":
-			self.infoLog.append("INF#003")
+		testLinkFunctions(self)
 	def isEmpty(self): return self.infoLog == [] and self.errorLog == []
 
-class Error:
-	def __init__(self):
-		self.errors = {}
-		self.infos = {"INF#001":"INFORMATION: Function normalize_url gave an invalid return value.",
-		"INF#002":"INFORMATION: Function github_repo_to_api gave an invalid return value.",
-		"INF#003":"INFORMATION: Function github_repo_to_api_releases gave an invalid return value."}
-	def getErrorString(self, errorCode): return self.errors[errorCode]
-	def getInfoString(self, infoCode): return self.infos[infoCode]
+def testLinkFunctions(test_function):
+	if normalize_url("/hallo.py.git") != "hallo.py":
+		test_function.infoLog.append("INF#001")
+	if github_repo_to_api("https://github.com/mitmirzutun") != "https://api.github.com/repos/mitmirzutun":
+		test_function.infoLog.append("INF#002")
+	if github_repo_to_api_releases("https://github.com/mitmirzutun") != "https://api.github.com/repos/mitmirzutun/releases":
+		test_function.infoLog.append("INF#003")
+
+def testNormalizeVersions(test_function):
+	if normalize_version("","") != "" or normalize_version("v1.-1") != "1.1":
+		test_function.infoLog.append("INF#004")
 
 def runTest(run = "run"):
 	if run == "run":
@@ -43,4 +44,4 @@ def runTest(run = "run"):
 		print("Test run surpressed.")
 
 if __name__ == "__main__":
-	runTest("surpress")
+	runTest()
